HoverBehaviour
Implemented the highlighting of objects when the mouse hovers over a
gameobject.
There is an event called when the mouse hovers over a gameobject and
when the mouse exits the object it sets the outline of the object back
to its default color.
To see the implementation work you must change the camera from
orthographic to perspective.