Worked on assessment build
Fixed the glitching of the tool tip when hovering over resources,
however, the tool tip will be displayed through other opened menus
unless the user turns the tool tip off in the options menu. Added
functionality for the camera to zoom in to a unit or enemy that has been
double clicked on. Updated the prefabs to not have mesh colliders.