Bug Fix Update
The hotkeys now work accordingly for the Action, Crafting and Units
tabs. The sclaing works correct. Fixed an issue causing weird look
rotations for the units. Updated the UIWorkshop script to handle the
functionality of  aprevious test script. Fixed a bug that was preventing
the User from crafting the steel resource. Added functionality to the
UntController script that when the user uses the Harvest button, if no
unit is selected it will select every unit and send them to harvest, or
if a unit or units are selected it will only send the unit(s) selected
to harvest. Fixed the bug that was cauusing issues with animations not
working correctly for the units.